requirements.txt
Flask == 3.0.*
Werkzeug == 3.0.*
requests == 2.31.* 
Flask-Caching == 1.10.*
gunicorn == 21.2.*
mysql-connector-python == 8.3.*